[brain tests] Burst threading from the main file
